+++ src/main/java/com/KookBee/portfolioservice/controller/NoteController.java
@@ -3,7 +3,7 @@ package com.KookBee.portfolioservice.controller;
 import com.KookBee.portfolioservice.domain.entity.Note;
 import com.KookBee.portfolioservice.domain.request.NoteCreateRequest;
 import com.KookBee.portfolioservice.domain.request.NoteEditRequest;
-import com.KookBee.portfolioservice.domain.response.CurriculumListResponse;
+import com.KookBee.portfolioservice.domain.response.studyResponse.CurriculumListResponse;
 import com.KookBee.portfolioservice.domain.response.noteResponse.NoteListResponse;
 import com.KookBee.portfolioservice.service.NoteService;
 import lombok.RequiredArgsConstructor;
