--- src/main/java/com/KookBee/portfolioservice/domain/response/PortfolioStudyCheckResponse.java
@@ -1,16 +1,0 @@
-package com.KookBee.portfolioservice.domain.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.util.List;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter @Setter
-public class PortfolioStudyCheckResponse {
-    private boolean isStudyJoin;
-    private List<PortfolioStudyLectureResponse> responseList;
-}
