+++ src/main/java/com/KookBee/portfolioservice/service/NoteService.java
@@ -6,7 +6,7 @@ import com.KookBee.portfolioservice.client.UserServiceClient;
 import com.KookBee.portfolioservice.domain.entity.Note;
 import com.KookBee.portfolioservice.domain.request.NoteCreateRequest;
 import com.KookBee.portfolioservice.domain.request.NoteEditRequest;
-import com.KookBee.portfolioservice.domain.response.CurriculumListResponse;
+import com.KookBee.portfolioservice.domain.response.studyResponse.CurriculumListResponse;
 import com.KookBee.portfolioservice.domain.response.noteResponse.NoteListResponse;
 import com.KookBee.portfolioservice.repository.NoteRepository;
 import com.KookBee.portfolioservice.security.JwtService;
